// frontend/src/entities/nutrtition/model/type.ts
export interface NutritionDay {
    id: number;
    publicId: string;
    name: string;
    date: string;
    description?: string;
    calories: string;
    protein: string;
    fat: string;
    carb: string;
}

export interface NutritionForm {
    name: string;
    description: string;
    calories: string;
    protein: string;
    fat: string;
    carb: string;
    dayDate: string;
}